refactor: CarName 객체 분리
